nnproj-learn/requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astunparse==1.6.3
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
imagecodecs==2020.2.18
imageio==2.8.0
ipaddr==2.2.0
joblib==0.15.1
keract==3.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lockfile==0.12.2
Markdown==3.2.2
matplotlib==3.2.1
msgpack==0.6.2
networkx==2.4
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.0.0
tifffile==2020.5.11
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
